+++ src/__tests__/Actions-test.js
@@ -66,28 +66,6 @@ describe('Actions', () => {
       expect(dispatch.firstCall.args[1]).to.deep.equal({ foo: 'bar' });
     });
 
-    it('warns if actions have not been added to a Flux instance', () => {
-      const actions = new TestActions();
-      const warn = sinon.spy(console, 'warn');
-
-      actions.getFoo();
-
-      expect(warn.firstCall.args[0]).to.equal(
-        'You\'ve attempted to perform the action TestActions#getFoo, but it '
-      + 'hasn\'t been added to a Flux instance.'
-      );
-
-      actions.asyncAction();
-
-      expect(warn.secondCall.args[0]).to.equal(
-        `You've attempted to perform the asynchronous action `
-      + `TestActions#asyncAction, but it hasn't been added `
-      + `to a Flux instance.`
-      );
-
-      console.warn.restore();
-    });
-
     it('sends return value to Flux dispatch', () => {
       const actions = new TestActions();
       const actionId = actions.getActionIds().getFoo;